+++ package.json
@@ -4,7 +4,7 @@
   "description": "o_O\"",
   "main": "index.js",
   "scripts": {
-    "hotbuild": "./node_modules/.bin/webpack --config webpack.config.js --watch",
+    "hotbuild": "./node_modules/.bin/webpack --progress --debug --watch --config webpack.config.js --display-chunks",
     "start": "node src/generate/app.bundle.js"
   },
   "repository": {
