+++ src/model/apiAction.js
@@ -1,0 +1,3 @@
+export const GET_BALANCE = 'getBalance';
+export const GET_MARGIN_STATUS = 'getcollateral';
+export const GET_HISTORY = 'getexecutions';
