--- src/model/apiType.js
@@ -1,5 +1,0 @@
-const apiType = {
-  getBalance: 'getBalance',
-  getMarginStatus: 'getcollateral',
-  getHistory: 'getexecutions'
-}
